tests(migrations) unit tests for core/009 (pg only)
Note that a problem was found in the migration while writing these
tests:

When the migration had already done the "up" action but had yet to do
the "teardown", an old node could create entities with NULL in the
ws_ids (this should not be allowed, they should have the
default ws_id instead). The fix involved moving the reset of the
default for that column to the teardown part of the migration.
(At which point, old nodes should already be shut down)